Third Commit
Added: file viewer, post link to homepage, authentication, file download to local system, post pagination, admin dashboard.

+++ app/Http/Controllers/PostsController.php
@@ -3,28 +3,41 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage; //lib for public folder storage
 use App\Post; //fetch post
 use DB; //to not use eloquent but sql query
+use Input;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct() //this function blocks off posts from unauthenticated users
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]); //except pages index and show, so guest can view posts
+    }
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
-        //$posts = Post::all(); [return all post data in table]
-        //return Post::where('title', 'Post Two')->get();  [fetching single post by title]
-        //$posts = DB::select('SELECT * FROM posts'); [using sql query instead of eloquent]
-        //return Post::orderBy('title', 'desc')->take(1)->get();
-        //$posts = Post::orderBy('title', 'desc')->get(); //fetches data in the table by title in descending order
-        
+    public function index(Request $request)
+    {  
         $posts = Post::orderBy('created_at', 'desc')->paginate(10);
-        return view('posts.index')->with('posts', $posts); //returns view of index page in posts folder
+        
+ 
+    return view('posts.index')->with('posts', $posts);
+      
+        $posts->appends(Request::query())->render();
+
         
     }
+       
 
     /**
      * Show the form for creating a new resource.
@@ -34,6 +47,7 @@ class PostsController extends Controller
     public function create()
     {
         return view('posts.create');
+
     }
 
     /**
@@ -46,14 +60,35 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'issn' => 'nullable',
+            'cover_image' => 'nullable|max:9999'
+           
         ]);
+        //handle file upload
+        if($request->hasFile('cover_image')){
+            //get filename with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }else{
+            $fileNameToStore = 'noimage.jpg';
+        }
         //create post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->issn = $request->input('issn');
+        $post->user_id = auth()->user()->id; //gets current user id and saves it to post id field
+        $post->cover_image = $fileNameToStore;
         $post->save();
-
+    
         return redirect('/posts')->with('success', 'Post Created');
     }
 
@@ -79,6 +114,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorised page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -93,12 +133,31 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'issn' => 'nullable'
         ]);
+
+        if($request->hasFile('cover_image')){
+            //get filename with extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+        
         //update edited post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        //test to see if user actually uploaded file
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -115,7 +174,19 @@ class PostsController extends Controller
     {
         //deletes posts in db
         $post = Post::find($id);
+
+        //check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorised page');
+        }
+        if($post->cover_image != 'noimage.jpg'){
+            //delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
+
+    
 }
